Split GateOpMixin into GroupMixin and AdjointMixin
